organize icon text comp
